feat: 특정 채팅방의 채팅 내역들 db에서 가져오기 기능 수현

+++ backend/src/main/java/com/achiko/backend/repository/ChatParticipantRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.achiko.backend.entity.ChatParticipantEntity;
 import com.achiko.backend.entity.UserEntity;
 
-public interface ChatParticipantRepository extends JpaRepository<ChatParticipantEntity, Integer> {
+public interface ChatParticipantRepository extends JpaRepository<ChatParticipantEntity, Long> {
 
 	List<ChatParticipantEntity> findByHost_UserId(Long hostId);
